\name{EuclideanDijkstra}

\alias{EuclideanDijkstra-package}
\docType{package}
\title{
\packageTitle{EuclideanDijkstra}
}
\description{
\packageDescription{EuclideanDijkstra}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EuclideanDijkstra}
\packageIndices{EuclideanDijkstra}

}
\author{
\packageAuthor{EuclideanDijkstra}

Maintainer: \packageMaintainer{EuclideanDijkstra}
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}

\keyword{ package }
\seealso{
\url{https://github.com/henkar91/AdvRprogr_lab3}

}
\examples{
euclidean(1000, 10)
dijkstra(wiki_graph, 1)
}
